Enhance axis and link management functionality
- Updated axis management functions to improve type safety and memory handling, including changes to `create_axis`, `delete_axis`, and related functions to use `ushort` and `uint` types.
- Implemented `delete_link` functionality to allow removal of links between nodes, including necessary memory adjustments and file synchronization.
- Added command handlers for `delete-link` and `print-free-space` to the CLI, enhancing user interaction and providing detailed free space information.
- Improved Makefile to dynamically include source and header files, streamlining the build process.
- Updated documentation to reflect new features and usage examples for axis and link management operations.

+++ src/axis.h
@@ -14,10 +14,10 @@
 #define AXIS_TIME 3
 
 // Function declarations
-int get_axis_count(uchar* node, int channel_index);
-int get_axis_offset(uchar* node, int channel_index, int axis_number);
-bool has_axis(uchar* node, uint channel_offset, int axis_number);
-int create_axis(int node_index, int channel_index, int axis_number);
-int delete_axis(int node_index, int channel_index, int axis_number);
+int get_axis_count(uchar* node, ushort channel_index);
+int get_axis_offset(uchar* node, ushort channel_index, ushort axis_number);
+bool has_axis(uchar* node, uint channel_offset, ushort axis_number);
+int create_axis(uint node_index, ushort channel_index, ushort axis_number);
+int delete_axis(uint node_index, ushort channel_index, ushort axis_number);
 
 #endif
